DEV: Flatten perf formatter to a single class and drop its unit specs
The performance formatter is developer tooling, so simplify it. Collapse the
former Discourse::PerformanceFormatter (and its Capture / RequestGroup /
Serializer modules) into a single top-level RspecPerformanceFormatter whose
helpers are class methods. It stays in spec/support (it is rspec-only);
bin/turbo_rspec requires it so the turbo parent can resolve
--format RspecPerformanceFormatter.

Drop the formatter's own unit specs (capture / request-group / serializer /
request) — it is manually verified dev tooling. Keep spec/lib/method_profiler_spec.rb,
which covers the one change that touches a real lib: MethodProfiler's opt-in
itemization.
