+++ spec/support/perf_capture.rb
@@ -2,24 +2,24 @@
 
 RSpec.configure do |config|
   config.before(:suite) do
-    if Discourse::PerformanceFormatter.enabled?
+    if RspecPerformanceFormatter.enabled?
       MethodProfiler.ensure_discourse_instrumentation!
       MethodProfiler.itemize_enabled = true
     end
   end
 
   config.around(:each) do |example|
-    unless Discourse::PerformanceFormatter.enabled?
+    unless RspecPerformanceFormatter.enabled?
       example.run
       next
     end
 
     test_summary = nil
     request_groups =
-      Discourse::PerformanceFormatter::Capture.collect_requests do
-        test_summary = Discourse::PerformanceFormatter::Capture.measure { example.run }
+      RspecPerformanceFormatter.collect_requests do
+        test_summary = RspecPerformanceFormatter.measure { example.run }
       end
 
-    example.metadata[:perf] = Discourse::PerformanceFormatter.assemble(test_summary, request_groups)
+    example.metadata[:perf] = RspecPerformanceFormatter.assemble(test_summary, request_groups)
   end
 end
